Remove the 'e' from the word 'Archive' of the project name
I had seen the missing 'e' in the word 'Archiv' as a spelling mistake
and inserted the 'e' in the new building of the repository.

With this commit I undo this change, because the archive should be a
rebirth of the PureBasic CodeArchiv from the www.purearea.net website
and I want to keep the name of the archive from there.