PrintSelection: Fix that the tool does not exit in some cases
One user of the tool reported that when he cancels PDF printing, he
normally receives a MessageBox telling him that he has canceled the
print operation. This MessageBox is not displayed when Notepad is
started hidden, which prevents him from confirming the MessageBox with a
click on the OK button, causing programs to wait endlessly in the
background until the OK button is pressed. Further calls of the tool
thus create more and more programs running in the background.

Notepad is therefore now no longer started hidden, so that the user
can react to the MessageBox.

Notepad version: 11.2206.17.0
Windows version: 11 Home 21H2